// src/modules/Artifacts/types/dto/persona-update.ts
export interface PersonaUpdateDto {
  name: string;
  age: number;
  gender: string;
  language: string;
  supportLevel: string;
  model: string;
  about: string;
  interaction: string[];
  cognition: string[];
  communication: string[];
  behavior: string[];
  stressfulActivities: string[];
  calmingActivities: string[];
  stereotypes: string[];
  softwareAspects: string[];
  socialAspects: string[];
}